Extracted extensions that are related to connecting to a json based model
