+++ CsCore/CsCoreUnity/Plugins/CsCoreUnity/com/csutil/ui/mtvmtv/ViewModelJsonExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7e54e1060b4a1648af811514a85191a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
